+++ inc/itilfollowup.class.php
@@ -375,7 +375,9 @@ class ITILFollowup  extends CommonDBChild {
       $itemtype = $input['itemtype'];
       $input['timeline_position'] = $itemtype::getTimelinePosition($input["items_id"], $this->getType(), $input["users_id"]);
 
-      $input["date"] = $_SESSION["glpi_currenttime"];
+      if (!isset($input['date'])) {
+         $input["date"] = $_SESSION["glpi_currenttime"];
+      }
       return $input;
    }